model 7 scenarios (#60)
    Category: scenarios
    JIRA issue: MIC-4539
    Research reference: scenarios

Changes and notes
Define scenarios.
Use scenario to determine exposure values for MAM, SAM, and SQ-LNS treatment.
Define index of data in loader function. This should have made it in the artifact PR but it somehow didn't. The artifact I was testing model 7 with had this index update but when I remade this key using the model_7_develop there was no index. I'm adding it here to avoid opening a new PR.

Verification and Testing
In an interactive sim, checked that the baseline scenario had mixed exposures for MAM, SAM, and and none for SQ-LNS (even if age-eligible). Checked that SQ-LNS scenario had no coverage for MAM and SAM and full exposure for SQ-LNS for simulants of eligible age.